+++ src/test/java/com/pivovarit/collectors/parallelToOrderedStream/ParallelOrderedBDDTest.java
@@ -10,8 +10,11 @@ import java.util.List;
 import java.util.Random;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
+import java.util.stream.Stream;
 
 import static com.pivovarit.collectors.ParallelCollectors.parallelToOrderedStream;
+import static com.pivovarit.collectors.ParallelCollectors.parallelToStream;
+import static com.pivovarit.collectors.infrastructure.TestUtils.TRIALS;
 import static com.pivovarit.collectors.infrastructure.TestUtils.returnWithDelay;
 import static java.time.Duration.ofMillis;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -32,7 +35,18 @@ public class ParallelOrderedBDDTest extends ExecutorAwareTest {
           .collect(Collectors.toList());
 
         assertThat(result).isSorted();
+    }
 
-        executor.shutdownNow();
+    @Property(trials = TRIALS)
+    public void shouldCollectToListInCompletionOrder() {
+        // given
+        executor = threadPoolExecutor(4);
+
+        List<Integer> result = Stream.of(350, 200, 0, 400)
+          .collect(parallelToStream(i -> returnWithDelay(i, ofMillis(i)), executor, 4))
+          .limit(2)
+          .collect(Collectors.toList());
+
+        assertThat(result).isSorted();
     }
 }
